Atualizando script de empacotamento.

+++ trunk/build.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 OLDDIR=$PWD
-PACKAGE=`basename $OLDDIR`
+PACKAGE=g3270
 
 RPMDIR=`rpm --eval="%{u2p:%{_rpmdir}}"`
 RPMARCH=`rpm --eval="%{u2p:%{_target_cpu}}"`
@@ -13,17 +13,28 @@ mv *~ /tmp
 
 make -C src clean
 
-cd ..
-tar -zcvf `rpm --eval="%{u2p:%{_sourcedir}}"`/$PACKAGE.tar.gz --exclude=*.svn* $PACKAGE
-
-if [ "$?" != "0" ]; then
-   exit -1
-fi
-
 cd $OLDDIR
 
 rm -f $RPMDIR/$RPMARCH/$PACKAGE*.rpm
 rm -f $RPMDIR/noarch/$PACKAGE*.rpm
+mv -f $SRCDIR/$PACKAGE*.rpm $HOME
+mv -f $SOURCES/$PACKAGE* $HOME
+
+rm -fr /tmp/$PACKAGE
+mkdir -p /tmp/$PACKAGE
+if [ "$?" != "0" ]; then
+   exit -1
+fi
+
+cp -R * /tmp/$PACKAGE
+if [ "$?" != "0" ]; then
+   exit -1
+fi
+
+tar -zcvf `rpm --eval="%{u2p:%{_sourcedir}}"`/$PACKAGE.tar.gz --directory=/tmp --exclude=*.svn* $PACKAGE
+if [ "$?" != "0" ]; then
+   exit -1
+fi
 
 rpmbuild -ba $PACKAGE.spec
 if [ "$?" != "0" ]; then
@@ -35,15 +46,19 @@ PACKET_VERSION=`rpm -qp --qf '%{VERSION}-%{RELEASE}' $SOURCE_PACKET`
 
 if [ "$1" != "--debug" ]; then
    echo Gerando copia do repositorio para a tag $PACKET_VERSION
-   svn rm --force -m "Atualizando $PACKET_VERSION" http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/tags/$PACKET_VERSION
-   svn cp -m "Pacote $PACKET_VERSION" http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/trunk http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/tags/$PACKET_VERSION
 
-   echo "Enviando arquivo source para o servidor..."
-   scp `rpm --eval="%{u2p:%{_srcrpmdir}}"`/$PACKAGE*.src.rpm $USER@suportelinux.df.bb.com.br:/dados/src/$VENDOR
-   if [ "$?" != "0" ]; then
-      echo "Erro ao copiar o pacote fonte"
-      exit -1
+   if [ "$VENDOR" = "suse" ]; then
+   	svn rm --force -m "Atualizando $PACKET_VERSION" http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/tags/$PACKET_VERSION
+   	svn cp -m "Pacote $PACKET_VERSION" http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/trunk http://$USER@suportelinux.df.bb.com.br/svn/$PACKAGE/tags/$PACKET_VERSION
    fi
+
+#   echo "Enviando arquivo source para o servidor..."
+#   scp `rpm --eval="%{u2p:%{_srcrpmdir}}"`/$PACKAGE*.src.rpm $USER@suportelinux.df.bb.com.br:/dados/src/$VENDOR
+#   if [ "$?" != "0" ]; then
+#      echo "Erro ao copiar o pacote fonte"
+#      exit -1
+#   fi
+
    echo "Enviando arquivo binario o servidor..."
    scp $RPMDIR/$RPMARCH/$PACKAGE*.rpm $USER@os2team.df.intrabb.bb.com.br:/home/matriz/pacotes/$VENDOR/bb
    if [ "$?" != "0" ]; then
